# api/controllers/trades_controller.go
package controllers

import (
	//	"bytes"

	"encoding/json"
	"errors"
	"fmt"
	"io/ioutil"
	"log"
	"net/http"
	"strings"

	"strconv"

	helpers "github.com/ahmed-023/bitget-helpers"
	"github.com/kryptomind/bidboxapi/AccountsService/api/models"
	"github.com/kryptomind/bidboxapi/AccountsService/api/response"
)

type Data struct {
	MarginCoin        string `json:"marginCoin"`
	Symbol            string `json:"symbol"`
	HoldSide          string `json:"holdSide"`
	OpenDelegateCount string `json:"openDelegateCount"`
	Margin            string `json:"margin"`
	Available         string `json:"available"`
	Locked            string `json:"locked"`
	Total             string `json:"total"`
	Leverage          int    `json:"leverage"`
	AchievedProfits   string `json:"achievedProfits"`
	AverageOpenPrice  string `json:"averageOpenPrice"`
	MarginMode        string `json:"marginMode"`
	HoldMode          string `json:"holdMode"`
	UnrealizedPL      string `json:"unrealizedPL"`
	LiquidationPrice  string `json:"liquidationPrice"`
	KeepMarginRate    string `json:"keepMarginRate"`
	MarketPrice       string `json:"marketPrice"`
	CTime             string `json:"cTime"`
}

type Positions struct {
	Code        string `json:"code"`
	Msg         string `json:"msg"`
	RequestTime int    `json:"requestTime"`
	Data        []Data `json:"data"`
}

func getTradeData(api_key string, secret_key string, passphrase string) ([]Data, error) {
	host := "https://api.bitget.com"
	path := "/api/mix/v1/position/allPosition?productType=sumcbl"
	url := host + path
	method := "GET"

	client := &http.Client{}
	req, err := http.NewRequest(method, url, nil)

	if err != nil {
		return []Data{}, err
	}

	server_time := helpers.GetBitgetServerTimeStamp()
	signatures := helpers.GenerateBitgetSignature(secret_key, api_key, passphrase, "GET", path, server_time)

	req.Header.Add("ACCESS-KEY", api_key)
	req.Header.Add("ACCESS-SIGN", signatures)
	req.Header.Add("ACCESS-TIMESTAMP", server_time)
	req.Header.Add("ACCESS-PASSPHRASE", passphrase)
	req.Header.Add("Content-Type", "application/json")

	res, err := client.Do(req)
	if err != nil {
		return []Data{}, err
	}
	defer res.Body.Close()

	body, err := ioutil.ReadAll(res.Body)
	if err != nil {
		return []Data{}, err
	}

	// Parse the API response into a AllPositionsResponse struct
	var resp Positions
	if err := json.Unmarshal(body, &resp); err != nil {
		// handle error
		return []Data{}, err
	}

	return resp.Data, nil
}

func (server *Server) GetOpenTrades(w http.ResponseWriter, r *http.Request, email string) {

	service := r.URL.Query().Get("service")
	if service == "" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service is required"))
		return
	}

	//get keys by user id
	key := models.Key{}
	keys, err := key.FindKeysByEmail(server.DB, email, service)
	if err != nil {
		response.JSON(w, http.StatusBadRequest, errors.New("User not found"))
		return
	}

	var api_key string
	var secret_key string
	var passphrase string
	if strings.ToLower(service) != "bitget" {
		response.JSON(w, http.StatusNoContent, errors.New("Exchange coming soon"))
		return
	} else {
		api_key, err = helpers.DecryptStrings(keys.ApiKey)
		if err != nil {
			log.Fatal(err)
			return
		}
		secret_key, err = helpers.DecryptStrings(keys.SecretKey)
		if err != nil {
			log.Fatal(err)
			return
		}
		passphrase, err = helpers.DecryptStrings(keys.Passphrase)
		if err != nil {
			log.Fatal(err)
			return
		}
	}
	resp, err := getTradeData(api_key, secret_key, passphrase)

	if err != nil {
		log.Fatal(err)
	}

	var open_trades []Data
	// Iterate through each position in the response and print whether it's open or closed
	for _, pos := range resp {
		if pos.HoldSide == "long" || pos.HoldSide == "short" {
			if total, err := strconv.ParseFloat(pos.Total, 64); err == nil && total > 0 {
				open_trades = append(open_trades, pos)
			}
		}
	}

	json, err := json.Marshal(open_trades)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.Write(json)
}

func (server *Server) GetClosedTrades(w http.ResponseWriter, r *http.Request, email string) {

	service := r.URL.Query().Get("service")
	if service == "" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service is required"))
		return
	}

	//get keys by user id
	key := models.Key{}
	keys, err := key.FindKeysByEmail(server.DB, email, service)
	if err != nil {
		response.JSON(w, http.StatusBadRequest, errors.New("User not found"))
		return
	}

	var api_key string
	var secret_key string
	var passphrase string
	if strings.ToLower(service) != "bitget" {
		response.JSON(w, http.StatusNoContent, errors.New("Exchange coming soon"))
		return
	} else {
		api_key, err = helpers.DecryptStrings(keys.ApiKey)
		if err != nil {
			log.Fatal(err)
			return
		}
		secret_key, err = helpers.DecryptStrings(keys.SecretKey)
		if err != nil {
			log.Fatal(err)
			return
		}
		passphrase, err = helpers.DecryptStrings(keys.Passphrase)
		if err != nil {
			log.Fatal(err)
			return
		}
	}
	resp, err := getTradeData(api_key, secret_key, passphrase)

	var closed_trades []Data
	// Iterate through each position in the response and print whether it's open or closed
	for _, pos := range resp {
		if pos.HoldSide == "long" || pos.HoldSide == "short" {
			if total, err := strconv.ParseFloat(pos.Total, 64); err == nil && total > 0 {
				continue
			} else {
				closed_trades = append(closed_trades, pos)
			}
		}
	}

	json, err := json.Marshal(closed_trades)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.Write(json)
}

// Get User Orders godoc
// @Summary      Get User Orders
// @Description  Get User Orders
// @Tags         orders
// @Accept       json
// @Produce      json
// @Param        service  query  string true  "service"
// @Param        email  query  string true  "email"
// @Success      200  {object}  []models.Order
// @Failure      400  {string}  coin required
// @Router       /orders [get]
func (s *Server) GetUserOrders(w http.ResponseWriter, r *http.Request, email string) {

	service := r.URL.Query().Get("service")
	if service == "" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service required"))
		return
	}

	if service != "bitget" && service != "binance" && service != "bybit" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service not supported"))
		return
	}

	orders, error := models.GetOrdersByUserEmailAndExchange(s.DB, email, service)
	if error != nil {
		response.ERROR(w, http.StatusOK, errors.New("unable to get orders at the moment, try again later"))
		return
	}

	response.JSON(w, http.StatusOK, orders)
}

// Get User Open Positions godoc
// @Summary      Get User Open Positions
// @Description  Get User Open Positions
// @Tags         orders
// @Accept       json
// @Produce      json
// @Param        service  query  string true  "service"
// @Param        email  query  string true  "email"
// @Success      200  {object}  []models.Position
// @Failure      400  {string}  coin required
// @Router       /positions [get]
func (s *Server) GetUserOpenPositions(w http.ResponseWriter, r *http.Request, email string) {
	service := r.URL.Query().Get("service")
	if service == "" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service required"))
		return
	}

	if service != "bitget" && service != "binance" && service != "bybit" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service not supported"))
		return
	}

	var positions models.Position
	positionData, error := positions.GetOpenPositions(s.DB, email, service)
	if error != nil {
		response.ERROR(w, http.StatusOK, errors.New("unable to get orders at the moment, try again later"))
		return
	}
	response.JSON(w, http.StatusOK, positionData)
}

// Get User Cleared Hedge Positions godoc
// @Summary      Get User Cleared Hedge Positions
// @Description  Get User Cleared Hedge Positions
// @Tags         orders
// @Accept       json
// @Produce      json
// @Param        service  query  string true  "service"
// @Param        email  query  string true  "email"
// @Success      200  {object}  []models.GroupedPosition
// @Failure      400  {string}  coin required
// @Router       /hedge-positions [get]
func (s *Server) UserClearedHedgePositions(w http.ResponseWriter, r *http.Request, email string) {
	service := r.URL.Query().Get("service")
	if service == "" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service required"))
		return
	}

	if service != "bitget" && service != "binance" && service != "bybit" {
		response.ERROR(w, http.StatusBadRequest, errors.New("service not supported"))
		return
	}

	var positions models.Position
	positionData, error := positions.GetClosedGroupedPositions(s.DB, email, service)
	if error != nil {
		fmt.Println("🚀 ~ file: trades_controller.go:300 ~ func ~ error:", error)
		response.ERROR(w, http.StatusOK, errors.New("unable to get orders at the moment, try again later"))
		return
	}
	response.JSON(w, http.StatusOK, positionData)
}