# docs/swagger.yaml
basePath: /accounts
definitions:
...
  models.GroupedPosition:
    properties:
      hedgeID:
        type: string
      positions:
        items:
          $ref: '#/definitions/models.Position'
        type: array
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      email:
        type: string
      fee:
        type: number
      id:
        type: string
      margin_coin:
        type: string
      order_id:
        type: string
      order_price:
        type: string
      order_type:
        type: string
      position_id:
        type: integer
      profit:
        type: number
      quote_amount:
        type: string
      service:
        type: string
      side:
        type: string
      size:
        type: string
      symbol:
        type: string
    type: object
  models.Position:
    properties:
      created_at:
        type: string
      exchange:
        type: string
      fee:
        type: number
      first_buy_amount:
        type: string
      hedge_id:
        type: string
      id:
        type: integer
      last_update_price:
        type: string
      layer:
        type: integer
      leverage:
        type: string
      liq_price:
        type: string
      margin:
        type: string
      mark_price:
        type: string
      open_price:
        type: string
      order_id:
        type: string
      side:
        type: string
      size:
        type: string
      status:
        type: string
      stop_loss:
        type: string
      symbol:
        type: string
      take_profit:
        type: string
      total_profit:
        type: number
      unrealized_pl:
        type: string
      updated_at:
        type: string
      user_email:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is BidBox Accounts Service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BidBox Accounts Service API
  version: "1.0"
paths:
...
  /hedge-positions:
    get:
      consumes:
      - application/json
      description: Get User Cleared Hedge Positions
      parameters:
      - description: service
        in: query
        name: service
        required: true
        type: string
      - description: email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GroupedPosition'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get User Cleared Hedge Positions
      tags:
      - orders
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get User Orders
      parameters:
      - description: service
        in: query
        name: service
        required: true
        type: string
      - description: email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get User Orders
      tags:
      - orders
  /positions:
    get:
      consumes:
      - application/json
      description: Get User Open Positions
      parameters:
      - description: service
        in: query
        name: service
        required: true
        type: string
      - description: email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Position'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get User Open Positions
      tags:
      - orders
swagger: "2.0"